fix: entity 매핑 문제 수정

+++ src/main/java/com/juyb99/dinorunboot/dto/request/ScoreRequestDTO.java
@@ -1,4 +1,4 @@
 package com.juyb99.dinorunboot.dto.request;
 
-public record ScoreRequestDTO(String nickname, Long point) {
+public record ScoreRequestDTO(String nickname, Integer point) {
 }
